+++ src/main/java/nextstep/domain/repository/ReservationRepository.java
@@ -1,4 +1,4 @@
-package nextstep.repository;
+package nextstep.domain.repository;
 
 import nextstep.domain.Reservation;
